+++ lib/grruby.rb
@@ -1,5 +1,5 @@
 require_relative 'grruby/grruby'
 require_relative 'grruby/tasks/GrFunctionsClasses'
 require_relative 'rubyplot/PlotClasses/PlotClasses'
-require_relative 'rubyplot/scripting'
+require_relative 'rubyplot/figure'
 require_relative 'rubyplot/plotspace'
