+++ lib/rubyplot/plots.rb
@@ -1,0 +1,15 @@
+module Rubyplot
+  module Plots
+    class Base
+      include Rubyplot::GRWrapper::Tasks
+      def call
+        @tasks.each do |task|
+          task.call
+        end
+      end
+    end
+  end
+end
+
+require_relative 'plots/line'
+require_relative 'plots/scatter'
