+++ spec/scripting_layer/single_plot_graph_spec.rb
@@ -6,10 +6,10 @@ describe Rubyplot::Figure do
     @y1 = [10, 20, 30, 40, 50]
   end
 
-  context '#line' do
+  context '#line!' do
     it 'creates a simple line graph' do
       a = Rubyplot::Figure.new
-      a.line @x1, @y1
+      a.line! @x1, @y1
       a.save 'file_name.bmp'
 
       expect(equal_files('file_name.bmp', 'line_graph.bmp')).to eq(true)
@@ -17,17 +17,17 @@ describe Rubyplot::Figure do
 
     it 'creates a line graph with points marked' do
       a = Rubyplot::Figure.new
-      a.line @x1, @y1, markers: true
+      a.line! @x1, @y1, markers: true
       a.save 'file_name.bmp'
 
       expect(equal_files('file_name.bmp', 'line_points_graph.bmp')).to eq(true)
     end
   end
 
-  context '#scatter' do
+  context '#scatter!' do
     it 'creates a simple scatter graph' do
       a = Rubyplot::Figure.new
-      a.scatter @x1, @y1
+      a.scatter! @x1, @y1
       a.save 'file_name.bmp'
 
       expect(equal_files('file_name.bmp', 'scatter_graph.bmp')).to eq(true)
